Add some skew to our queries for calendar events
Looking a few days ahead and behind loads a little bit more data than
we need but insures we don’t get into any timezone trouble.

Adding 6 days to the start and end of a month allows us to populate
days that appear on the calendar like the 30th before the 1st.