Upgrade to ember simple auth 1.0
This is a big change in a couple of ways:
There are deprecations aplenty and a big API change that requires the
session to be an injected service.

Changed all the authenticated routes to use the new correct mixing

Changed the application route to override sessionInvalidated instead of
the sessionInvalidationSucceeded action.

Also had to write a custom ajax service to ensure that authentication
headers get sent with requests.

Had to modify the application adapter to send auth data with ember-data
requests.