revert to using s3-index deployment
Redis is a bust, it is completely insecure and making it secure is a
huge pain.  Instead lets ship the JSON file to S3 and pull it down over
SSL.   Placed all of the index assets into a single bucket with a
prefix that allows it to be versioned and separated by environment.