Improved loading of navigation
Has a loading indicator and does not load until permissions are
fulfilled.  This should make it easier to tell when there is a network
issue from the users perspective instead of just loading the dashboard
link and nothing else.

Remove navigation controller which as far as I can tell was never used.