Improved loading of navigation (#1555)
* Improved loading of navigation

Has a loading indicator and does not load until permissions are
fulfilled.  This should make it easier to tell when there is a network
issue from the users perspective instead of just loading the dashboard
link and nothing else.

Remove navigation controller which as far as I can tell was never used.

* Collapse if blocks into single block for clarity

These are all controlled by the same permissions so it makes sense for
them to be in the same block.