Use firefox in a framebuffer to run tests on travis (#1589)
* Use firefox in a framebuffer to run tests on travis

Hopefully this is more reliable that phantomjs.  Since we can run it in
a linux container we should get more tests to run and not having to
download travis or install nodejs should be much faster as well.

* Trigger keyup instead of search event

Firefox doesn't respond to the search event so we have to use keyup for
tests to pass in that browser.