Users can be moved into the current group
Allows users to be moved one at a time out of another group and into the
current one.  By not passing the learner group and by removing computed
properties and using a task instead the component doesn't re-render as
soon as the learner groups are modified providing a much improved user
experience.