Replace styles on new-user form
This is a new structured approach to styling components.  For now I'm
placing the files in newcomponents/ directory but once we have replaced
styles on the rest of the site the components/ directors will be
removed.

Individual ember components get a class and then we use mixins as
functions to apply the actual styles to the page.  Hopefully this gives
us a bunch of re-uability without needing to put things in a global
style space at all.