Remove flaky input focus test
This test fails in lots of cases where the browser chooses not to place
focus on the put.  For example if any other test has failed then focus
is on the failing test and can't be re-directed.  Or if testing in two
browsers this test will fail because one browser is is not focuses by
the OS.