Show a warning when API versions do not match
If a frontend attempts to contact an API with a mismatched version
string the user will see a message an be unable to continue. This will
prevent any kind of weird situation where an incorrect version gets
shipped or an advanced setup is set to point to an older API install.