Provide Notification When Program Is Deleted #2348
No confirmation message was provided for program delete.  Furthermore, the previous message was retained after delete.  As a result, a delete immediately after a save action provided an edit link to a deleted program, which caused null object errors. To fix, create a notification for program delete and clear the notification for saved program.

* update programs controller to clear save notification and add delete notification on program removal.

* update programs template to display removal notification on program removal.

* create text and translations for program removal.

* update tests to verify presence of removal notification

Provide Notification When Program Is Deleted #2348