Watch out for errors when building stewards list
Sometimes in the process of building the list of stewards the school
doesn't resolve.  I'm really not sure why this happens other than it
appears to be a temporary (possibly race) condition.  We can mitigate
all effects by just checking for the school the same way we check for
the department.