Check our addons for conflicting dependencies
When addons all depend on the same addon it can cause mysterious issues
as their output has the possibility of overriding each other.  This
ensures that when we add or update and addon that we test for this
problem.  The only conflict we currently have is with
ember-getowner-polyfill - I excluded it from our currently check and we
can work to bring versions into line through PRs.