Use Yarn for package management
Yarn is deterministic so we will always get the same set of packages
when running an install.  It is also cross compatible with NPM so we
can make this transition gradually.

At this point I have setup travis to NOT use the lockfile.  This will
ensure that our automatic greenkeeper updates are being tested
correctly.