Fix sessions filter on list to wait for input
Debouncing this filter and caching the value locally ensures that every
typed character doesn't cause the table to re-render which is slow and
very intensive. Instead we record each char as it is typed so the input
field stays up to date, but we only send the value up to be processed if
there is a 250ms delay.