Cache frontend using service workers
After a lot of experimentation I don't think we're ready for offline
fallback of API requests. What we are ready for is aggressive caching
and background loading of assets. This change will cause the app to boot
much faster the second time it is used because all of the JS and CSS
will be pre-loaded by the browser.