Remove the ember-service-worker-index addon
This is a little bit too aggressive in what it returns as the
`index.html` page. It also traps requests to /Shibboleth.SSO stuff which
the browser needs to be allowed to visit. Since we can't know in advance
what the shib route could be for any school the best choice is just to
remove this addon for now.