Display notice when user is offline
The more our app works while disconnected the more important it will be
to display a warning to users when they are not online.  This listens
for the browser offline/online events and puts up a warning when a user
is offline. It also starts a timer to check the connection status again,
but this is mainly an empty gesture for a better experience.