Disable the Froala beacon
Every time the editor is displayed it sends a ping image request back to
the froala mothership. They were a little bit cagey about what this is
for and we are fully licensed to distribute the editor in our opensource
project so we're going to go ahead and disable this phone home
functionality.