Use computed properties instead of tasks where appropriate
Tasks are great for managing some concurency, but they don't manage
keeping themselves up to date when changes are made as well as CPs do.
For displaying this learner group data CPs are the way to go.

Also modified the learnergroup-list to use the hasMany meta property
instead of loading all of the users in advance.