Limit the length of the URL when we pre-load data
When there are hundreds of sessions in a course the `filter` parameters
we send to pre-load objectives and session types for sessions can get
quite long and eventually break the Apache 8K upper limit for URL size.
This breaks these up into several smaller requests which do not reach
this limit.