Cleanup some Cruft
The session afterModel hook was doing some weird promise stuff to grab
the session types which it
turns out we are not actually even using anymore so I removed the
entire block.
We are no longer using showBackToCourseListLink anywhere to hide this
link so there is no reason to set it.