Update to Font Awesome 5
Replace FA4 icons with their FA5 equivalents
Use official addon in place of ember-font-awesome addon
Default to the pro regular style, except for the navigation which is in
solid.

Default to the solid style icon
After group review this is our best icon choice and we can default to
it.
Fixup tests with bad references to icons
Our move to FA5 invalidated these selectors
Remove local ember-fontawesome dependency