Navigation improvements
Students should will never see the navigation so we need a completely
different layout in that case with no left hand bar and better spacing on
top.  While working on the navigation I also added some hover effects
and expanded the link to take up all the space so you can click in a
larger area.