Add automated accessibility checks
This will decorate our application with indications where accessibility
is failing and put pressure on us to make it better. Once we get some of
the basics under control we can add some automated checks to the
rendering tests as well.

Also does some automated work on route changes to focus screen readers
on the correct content.