Indicate that all sessions are expanding
In big courses it can take a while for all of the sessions to open. What
I'm doing here is starting a loading indicator and then opening the
sessions. The browser takes a while to render the newly exposed DOM so
we use Ember.run.next() to wait for control to come back from that
process and then kill the spinner.