Handle icons in the froala editor
We don't have much control over this UI component. I was able to set the
icon template to work with font awesome 5, but then I needed to add a
method from the fontawesome library to convert the `<i>` tags that the
editor created into the SVG icons we can actually display.