Improve experience when the API version is wrong
Users sometimes see this screen if the service worker doesn't update
before the component renders. I've added a few seconds of delay before
the warning is shows, but in case a user still sees this overlay I've
also improved the experience to start an auto countdown and re-load when
the new worker is ready.