Fix tests now that icon title works
The title for SVG icons is in an element within the SVG not and
attribute of the icon. This caused some confusion for our tests both
when they look for the icon and when they grab the contents of a DOM
element (since title is now part of that text).