Improve test waiting for api-version-check
All of my attempts to get this test to wait for the task to finish have
load to flaky tests. Instead this change defines a property that will be
set on the UI when the task is done and then the test waits for that
property to show up before running the test.