Change the source order to improve loading performance
Our loader animation plays in the content for body and it contains it's
own style information. We want it to load and start rendering very
quickly and then the styles can be loaded for the main site. Since that
won't render until the scripts are done being parsed there is no risk in
starting the loading of styles late.