Add setup-authentication helper for acceptance tests
Placing this folder in addond-test-support ensures that it is merged
into the consuming application and callable. I also made a change that I
see commonly in the ember comunity to load this directory differently in
our internal dummy testing app to make it easy to use the helpers here
as well.