yanked this janky test.
converting this to a proper integation/rendering-test did not work - the
file upload field renders outside the test container. accessing the
component itself if verboten in this test context.
the only solution was to remove it completely, we just have to assume
that the underlying component library (ember-uploader) works as
intended. which is reasonable, IMO.