Manually specify asset path
Since we load our app and vendor.js files asynchronously using the defer
flag we need to pay special attention to the final path of our assets.
Otherwise ember-auto-import detects it automatically based on the last
script loaded (which for us is probably our custom build error, which is
wrong).