Add menuitem role to button
Our A11y checker flagged the local choose menubar because in its closed state it
doesn't contain any items. This is actually a menu with a submene, which
was mostly marked up correctly, just missing this one attribute to tell
screen readers that this button is an item in the menu.