Improve accessibility of links
More contrast was required between links and the background as well as
links and text they are embedded within. By making the standard text
slightly darker and modifying the shade of blue the contrast is
sufficient. When links are within text it is required that some visual
indication of focus be present, I've added underlines for that purpose.