Use fetch to make ember-data requests
Once we turn off jquery this will be the default, but for now I've
forced it on.

Changes for fetch based networking

With jQuery removed ember-data reverts to using fetch to get data, but
the simple auth mixin doesn't work for this so we have to set the
headers ourselves.