Created <Course::PublicationMenu> component
This extracts functionality from the shared PublishMenu component. That shared component requires a mixin on CourseHeader to work, so pulling this out into a single purpose component allows that mixin to be removed.

Needed to allow for a11y testing failures for the color contracst on scheduled courses which will have to be addressed at a later date.