Fix boolean expression in week-glance.js
The boolean expression checks .midnightAtTheStartOfThisWeek twice. One of
them was meant to be .midnightAtTheEndOfThisWeek instead.

Boolean OR was used where Boolean AND was meant. (If it is not midnight
at the begining of the week OR it is not midnight at the end of the
week, then the expression will always be TRUE. If AND is used instead,
it excludes midnight at the start or end of the week.)