Improve loading of course print view
Some larger courses were making 300+ network requests to load
the print view. This change optimizes the network loading in the same
way it is improved for the course editor and brings the number of
requests down into the 20-30 range.

I extracted this loading work into a utility so it's now shared between
these two routes.