Improve Session Overview Loading
All of the individual promises getting read in the load method was
causing a significant rendering delay as each one moved in and out of
the main thread. Loading them all at the same time solves that issue. As
long as the load has run at least once we should display the data. Also
moved one method from a task to a plain async method which it seems to
be.