Octaniefy <SessionCopy> component
I removed the validations here because they would never fire. Since the
save button is disabled when coruse or year isn't selected it wasn't
possible to ever be in an invalid state other than when there are no
courses in the year and that is represented in the UI bay "None" courses
and a disabled save button.