Replace save status modal with a live button
Because WaitSaving uses a modal dialog that gets wormholed out to a
different location in the DOM it is susceptible to missing it's queue to
be destroyed when the component containing it is removed. Instead of a
modal this makes the button into their own live saving state with a
spinner, percentage, and check mark when complete.