Add brotli compressed assets to the deploy
This modifies our deploy to now compress everything with both gzip and
brotli. Brotli is significantly more efficient and supported in most
browsers so we wand to use it wherever possible. However it doesn't work
at all over gzip and neither will work well if the correct headers are
not sent. So instead of overwriting the original file with a compressed
version this provided a .gz and .br version of each and we'll modify the
API server to decide which to send.