Pre Load Course Data
Eliminates up to 75 network requests in favor of loading all course data in one request and all data for sessions attached to the course in a second request. Breaking it up in two peices allows us to show the results from the course load while the sessions are still loading making the page feel more responsive.