Improve Courese Loading Experience
Placing loading routes for both course and course-index allows the two sections to animate in indeividually. The Course::SessionsLoading component is identical to the CourseSessions component, but I wasn't able to get CourseSessions to both wait for the data loader to complete and respond to change in the session list.