Load less data on learner group list
I've removed the course count and moved the check for linked courses
into the delete dialog. This way we don't need to load the entire
curriculum when loading the learner groups list.

To make the course information easy to get to it has been highlighted
and re-styled in the single group view.