Handle pre-loading of session data in CourseSessionsComponent
There is a bug in the ember loading route that causes query params to be
dropped so we can't use it here. Instead by annotating the getter for
sessionsCount and then passing this into did-update we can track when
the number of sessions change and ensure the load method is fired again
while re-using the existing loading state of this component.