Octanefiy DashboardMyreports component
Removed the report title component because it wasn't necessary anymore,
we have to generate all of the titles during the initial display in
order to sort correctly. The original sorting wasn't actually working as
reports with a default empty title would be pushed to the top, it just
looked ok because most of them start with `All...`.