Refactor config attributes
Only updating when the school changes means we'll miss any new config
relationships that get added, instead we track the configurations
relationship. This won't result in an update when a config value changes
though so we track the config model and then use a getter (which will
auto track and update) in order to extract the value.