Use resources to resolve async values
Instead of doing this work in a task I've created two generic reusable
resources. Resolve Async Value will resolve a promise and return the
resolved value, Async Process is passed a function and arguments and can
use that to do complicated async work.  I've then put the accessors for
these values into a getter to handle edge case processing and ensuring
the preloader is called.