Add route announcer
This a11y improvement will likely be making it into the default
blueprint soon, but we can add it now for improved accessibility for
screen readers. When a new route is loaded this will read the page title
and then a translated `is loaded` message to give an indication that
following a link changed the page.