Give the events grid some columns
We use span 50 and fractions of that to place events on the calendar
along side other events at the same time. This mostly worked, but
without specifying the number of columns in the grid sometimes it
overflows on mobile. This fix uses 50 evenly spaced columns to fill out
each day of the calendar.