corrects state management for creating new session types.
this was buggy before.
after creating a new session type, the component was collapsed. upon
exanding, the user was shown the opened form for creating a new session
type again.
this corrects this problem - the user is now shown the expanded list of
session types (including the newly one created).