Refactor course sessions
Extracted the session object into it's own class so we can use resources
and getters to fetch and display data. Modernized the component a bit to
use resources instead of a loader. Both of these changes allow us to
update ED versions as the firstOfferingDate CP on the session model
wasn't updating because it has too many levels of dependencies.