Remove mixin cruft on publication issues
All of the variables here existed to be able to use the same mixin for
all publishable items, but we got rid of that mixin a while ago. Instead
we can just check the values we need for a session and remove the extra
work. This has the added benifit of fixing an issue we were having with
upgrading ember data.