Add new release automation
This creates a manually runnable workflow to create a new git tag and
then automated the building of release notes from that tag. This will
allow us to simplify our chatops as only a new semver type will need to
be sent to Github and not other work will need to be done by the bot
server.