Drop ember-mockdate-shim
This is a super useful thing that we should probably use a lot more,
unfortuantly the ember addon isn't seeing any love and some long
standing PRs to bring it up to date remain open. We don't need a shim to
do the import anymore though so we can drop it. The real magic is in not
messing up backburner during a test run, but we can handle that in our
own helper here.