Improve flat map resource
Two edge cases here, one is that ember data is weird about map-by so we
have to resolve that first. More interestingly is that since we await
something before accessing the mapby we fail to entangle the individual
elements of each array. If we manually access those element at the top
it starts auto tracking and everything is peachy.