Invert session overriding in publish all
Because the optional and required publication issues may not be fully
resolved when the constructor runs it is necessary to turn
sessionsToOverride into a getter which will update as those values
coalesce. So I've created an array of sessions to skip overriding and
use that to inform the final getter.