Don't reload promises over and over
Putting these in a resource calls the findAll on every getter use.
Additionally because the school model is configured not to go to the
network every time if the current users' default school is loaded first
that will populate all the schools and nothing else will be loaded. This
isn't an issue in the full frontend, but breaks tests.