Using peer dependencies to help embroider build
For JS libs we have to let embroider know on the frontend
that common needs this stuff. Listing them as peer dependencies is the
way to expose that information during the build.

Ember addons have a process for doing this automatically and can be left
in dependencies without adding them as peer deps.