Fix pull request preview action
The `pull_request_target` event loads to a checkout of `master` and not
the PR commit. Since we're only running this action when a label has
been added we can go ahead and checkout the PR ref and run there. I also
added some more events so deployed PRs will get re-deployed when they
are changed.