Remove our 'preview' environment in favor of an ENV var
The preview environment was always a hack on our end and we end up
jumping through hoops to keep it working and use it in only one place
for our netlify builds. This actually causes our netlify builds to be
less like production as they end up including a bunch of code that Ember
normally strips out for production.