Fix annoying horizontal scroll
This appears to be a known browser bug, I'm not sure why it is popping
up for us now, but using vw units doesn't account for any scroll bars so
anytime there is a need for vertical scroll we were getting horizontal
bars as well. By using a 100% of the body we get the same effect without
the bug.