Make all peer dependencies *
Specifying a version for these seems to be mostly pointless as we're
only putting them here to let embroider know how to handle our build.
When something like dpeendabot does an update it doesn't automatically
widen this range to match so we end up out of sync. Going to try
allowing any version to see if that works long term.