Improve loading of a course
Load important course and session data so we don't need to make multiple
requests.

We can't use the course API endpoint to load the sessions because
ember-data halts the second request if the first is still in flight.
Instead we send a filter request to the API to get sessions attached to
the course.