Re-style top of course page
Removing the background color in favor of a small border at the bottom
of the course header. Had to move some things around as the text needs
to be flush aligned with the border or the publishing menu will have a
smaller gap between it's bottom and the border and cause the page to
look tilted.