Fix navigation on tablets
On a tablet sized screen in some cases for no particular reason that I
can tell the navigation will get huge when some items are selected.
Setting the body to 1fr seems to fix this so it may just be a browser
bug with double auto and our many hidden items in a collapsed navigation
view.