Pad the left most table column
Previously this was only done on the session because the table
background there is not the same as our background, however there are
lots of places where this is true so I've moved it to the default table
structure. This allows the left most columns text to not run up against
the edge of the table.