Replace Sentry Integration
As of v7 the Ember integration has been split out into a separate
package. I've updated our sentry setup to use the minimum set of
required config to see if the new integration handles some of our old
issues correctly, we'll need to adjust as we see real data come in.