Don't compensate for non array's when set to utilities
Because we use optional chaining in some places I was allowing non
arrays to come in and out of utilities, but let's not. As a temporary
improvement they return and empty array, but I think after we finish
this work in frontend we can even remove that and just use the
assertion.

I wasn't able to find a non-overwhelming way to expect an assertion to
fail in tests so I've marked these tests as todo temporarily.