Prefer Date to ISO string for dates
Moment was nice enough to let us get away with both, but it's too messy
and since luxon requires that we know in advance what we're parsing it's
gotta be fixed. I've deprecated ISO strings in favor of JS Date objects
in most places. User/School events come from the API directly and don't
go through a model transform, so in that case I've allowed ISO string
only. This is probably something we'll need to fix eventually, or maybe
not.