Remove disable prototype extensions addon
We've removed these extensions, I'm removing this safety check for now
in order to allow ember data to be updated. When prototype extensions is
enabled Ember.A() becomes a noop, but with this addon installed it
clobbers the RecordArray objects from ember-data. Wile we don't use
Ember.A ourselves it is used in ember-composable-helpers. This is a
temporary solution until Ember is updated to prevent it from clobbering
ED arrays, or ember-composable-helpers is updated to not use Ember.A().