Improve Login Experience
I've aligned our configuration with the frontend to remove token refresh
and updated the login itself to use the token we pass and not a
refreshed version. We still attempt to use the token to get a new token
to ensure the passed token is valid and when there are errors in that process they will now
display correctly.