Separate Subject Reports into Components
Each type of report will get it's own component which will be
responsible for loading the data required for itself. This removes the
need for a super complicated reporting service as well as branching
template logic that was confusing.

Added a subject report for courses as a first demonstration of the
concept.