Replace resolve async value resource
Remove these resources and replace them with the TrackedAsyncData
resource from ember-async data. This performs better than our version
and is on the paved path for ember. The declarations are move verbose
requiring a double getter, but this is an OK compromise for now. With
time we may be able to refactor some of these getters out and replace
them with the load template helper from this addon.