Replace ResolveAsyncValue in models
The TrackedAsyncData provides a better way to manage asynchronous state,
however it is more verbose and in order to be auto tracked it requires
each step in an async getter chain to be explicitly coded. I focused on
the testable output of the non-private getters for this change and I
think I've replaced our original resources with these new getters. Long
term we probably need to remove many of the public getters on models
because doing things this way where we can't know exactly where in an
async chain a change may happen makes for some verbose and difficult to
read code.