Fix school selection in new user creation
We were doing a bad job choosing the users primary school and a worse
job ensuring that school was available to pick from to create a new
user. This was resulting in users in some schools being unable to add
new users because the wrong ID was being sent for each new user.