rm scrolling feature from calendar views.
in the current implementation, the top navigation is pushed out of view
when the daily or weekly calendar is loaded on the dashboard.
this violates the first principle of Dave-Driven Design (TM) - "Don't
confuse the user".
Since we don't have the level of control over the scrolling behaviour
that would allow us to correct this issue, or at least i can't figure
out how to do this, our only other option is to remove scroll-into-view
functionality from the calendars altogether.