Remove optimization for saving learner groups
This method was designed to only modify, and only return, groupss which
were modified which is nice because it limits network requests when
modifying groups trees with lots of subgroups. This turns out to be a
bad idea however because users may open multiple tabs and modify the
group tree in each tab. Any user errors in this process can lead to
changes not being reflected back to the API because the data store in
the tab they have open does not realize a change to other groups is
necessary. I've removed that optimization here. It's less network
efficient, however it should result in consistent data being sent to the
API.