Improve deprecation handling with embroider macros
We don't want to inject this code into every application, only when
we're developing. I've also decided to throw on new deprecations. This
will require making updates to dependabot PRs when a new deprecation is
added, but it will also prevent us from stacking up a ton of
deprecations and having to track them down.