Run Percy Tests More Deliberately
We're running out of screenshots every month. Instead of running these
tests on every single PR they will no run when a new tag is created,
every few days on a schedule, when we label a PR, and when we manually
run the test. I've setup the transitive dependency update action to add
this label so they will run on that PR as well.

Because we're no longer running these on un-trusted forks I've also
modified the type of action so it has access to our secrets and we can
hide the percy token.