Pass data instead of a report
Creating a fake subject report and then running it without saving it
requires filtering out the ghost reports all over the place and is very
error prone. Once the report it run it has to be cleaned up and visiting
a link can prevent that cleanup from happening at all. Instead I've
broken open the report runners and utility methods to take the
properties of a report instead of the report itself. This is more
verbose, but unlocks running the reports (both saved and ephemeral) in
the same path.