Load icons from sprite file
Instead of using the fontawesome javascript process we can load our
icons out of the sprite file and reference them in components.
This requires a larger initial download, however the browser will cache that data.
I think we can play some
pre-loading games like we do for fonts to push the sprite files along
with the initial page request to improve this even more.

This allows us to drop all of the fontawesome addon and move away from
our team needing to work to maintain that addon ourselves and should
unblock our move to pnpm.