Catch missing model 404s
I'm not sure if this broke or was never working but we're not displaying
a nice error when a model is missing. This doesn't happen that often for
us, but when something is deleted we need to do better than or standard
error message because refreshing the page will just continue to be
broken.  Instead we now display our "Rats!" message with a link back to
the dashboard.