Modify the serializer to work better across timezones
We need to lie to our API and pretend everything is in UTC and then lie
to EmberData and pretend it's in local time because at the end of the
day we really want a date string here and not a date. This is the
smallest change that would get this working across different browsers
and timezones, but a bigger refactor of dates is probably in order.