Skip some pnpm publish checks
Pnpm, in the publish workflow, checks to be sure it is on master. This
doesn't work with our workflow because we use a tag. This issue has been
open with pnpm for more than a year and the only way to get around it is
to ignore the git checks here. That's not a big deal as we're in a fresh
checkout.