Augment our ICS token generating algorithm
The existing method doesn't work in Safari in contexts that aren't
secure. This includes our browser tests. Should we generate tokens in
insecure contexts? I'm not sure, but we have always allowed Ilios to run
without SSL, it's not a great idea, but we haven't, so far, been in the
business of requiring a secure connection. This generation of an ICS
token doesn't, in my mind, provide less security. In an insecure context
we'd already be passing this data around anyway. The method itself is
secure enough that the resulting token should not be guessable, which is
all that is required.