Fix tagging workflow
The pnpm version command doesn't exist. It forwards to npm for some
bizare reason. Since we're using pnpm workspaces and not the npm kind
this doesn't work for us. Instead we can use this nice tiny npm package
`update-version` to accomplish the same thing and modify the
package.json the way we want.