Clean up peer dependency on mirage
Maybe just a glitch in the way the package manager works, but I think
maybe embroider of ember-cli-mirage may also be missing with this on
purpose. In either case moving this into being strictly a peer of
common so it has to be provided by host apps seems to make embroider
much happier.