Upgrade deprecation handler
This was added to common as a way to remove the deprecation handler that
isn't being maintained and isn't ready for embroider. Bringing it over
to frontend here. This handler has one major difference from the
original in that it will throw when new deprecations are added, this
should work to our long term advantage as we'll be able to see the
failure in CI immediately and either handle the deprecation or silence it.