Move linting into workspaces
In order to clean up our dependency tree I needed to move all of the
linting back into the workspaces and out of the root. This will make
updating the ember apps easier, but we'll have to figure out how to
configure editors to read the config at each level.